+++ ipbandroidapi/src/main/java/com/progressterra/ipbandroidapi/api/iECommersCoreApi/models/shop/PaymentTokenRequest.kt
@@ -1,8 +1,6 @@
 package com.progressterra.ipbandroidapi.api.iECommersCoreApi.models.shop
 
-import java.math.BigDecimal
-
 internal class PaymentTokenRequest(
-    val amountPayment: BigDecimal,
+    val amountPayment: Double,
     val payment_token: String
 )
